pyproject.toml
[project]
name = "telegrambotcli"
version = "0.5.2"
description = "A utility for quickly generating projects on aiogram 3"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "AnonimPython", email = "moscow.retro@list.ru"}
]
keywords = ["aiogram", "cli", "generator", "telegram", "bot"]
dependencies = [
    "typer[all]",
    "sqlmodel",
    "aiogram>=3.0.0",
    "pydantic-settings",
    "psycopg2",
    "python-dotenv",
    "pydantic",
]

[project.scripts]
telegrambotcli = "aiogram_gen.main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
